image-service/requirements.txt
fastapi==0.115.5
uvicorn==0.34.0
python-multipart==0.0.20
pillow==11.0.0
pytest==8.3.4
httpx==0.28.1

